refactor: split LevelCanvas into partial-class files by flow
