+++ src/CtrDxEditor.Shared/Rendering/LevelCanvas.Input.cs
@@ -1,0 +1,460 @@
+using System;
+using System.Collections.Generic;
+using System.Globalization;
+using System.Linq;
+
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Input;
+using Avalonia.Interactivity;
+using Avalonia.Media;
+using Avalonia.Media.Imaging;
+
+using CtrDxEditor.Content;
+using CtrDxEditor.Core.Document;
+using CtrDxEditor.Core.Editing;
+using CtrDxEditor.Core.Geometry;
+
+namespace CtrDxEditor.Rendering
+{
+    /// <summary>Pointer, wheel, and touch input: hit-testing, dragging, and gesture handling.</summary>
+    public sealed partial class LevelCanvas
+    {
+        // Hovering / dragging the auto-catch radius ring, or a horizontal rail end/hook, uses a horizontal-
+        // resize cursor (col-resize); a vertical rail end/hook uses the vertical one.
+        // Cached, but created lazily rather than in the static constructor: eager creation would touch
+        // Avalonia's cursor factory at type load, which throws in the headless test host. Lazy .Value
+        // still allocates each cursor only once, so pointer-move hit-testing doesn't churn instances.
+        private static readonly Lazy<Cursor> LazyResizeCursor = new(() => new Cursor(StandardCursorType.SizeWestEast));
+
+        private static readonly Lazy<Cursor> LazyVResizeCursor = new(() => new Cursor(StandardCursorType.SizeNorthSouth));
+
+        private static Cursor ResizeCursor => LazyResizeCursor.Value;
+
+        private static Cursor VResizeCursor => LazyVResizeCursor.Value;
+
+        // Whether a level-space point sits on the selected grab's auto-catch radius ring, within a
+        // ~6px screen tolerance (converted to level units by the current zoom).
+        private bool OnRadiusEdge(Vec2 levelPt)
+        {
+            return SelectedObject is { } g && View.Zoom > 0 && RadiusRing.Of(g) is { } ring
+                && GrabRadius.OnEdge(new Vec2(g.X, g.Y), ring.Radius, levelPt, 6 / View.Zoom);
+        }
+
+        // What part of the selected movable grab's rail a level point is over, or None. The hit-testing
+        // itself lives in GrabRail; here we only supply the selected grab's geometry and the screen-derived
+        // tolerances: ~9 px for the end caps, the hook's own footprint, and the bar's half thickness.
+        private GrabRail.Handle HitRail(Vec2 levelPt)
+        {
+            return SelectedObject is { Type: "grab" } sel
+                && View.Zoom > 0
+                && GrabRenderer.DrawsMovableRail(sel)
+                && GrabRail.Of(sel) is { } g
+                ? GrabRail.HitTest(g, levelPt, endTolerance: 9 / View.Zoom, hookTolerance: 24, barThickness: 20)
+                : GrabRail.Handle.None;
+        }
+
+        // What part of the selected rotatable object's dial a level point is over, or None. The geometry
+        // lives in ObjectRotation; here we supply the object's spec and the screen-derived tolerances
+        // (converted to level units by the current zoom), matching the rail/radius handles.
+        private ObjectRotation.Handle HitRotationDial(Vec2 levelPt)
+        {
+            if (SelectedObject is not { } obj || View.Zoom <= 0 || RotationTable.For(obj.Type) is not { } spec)
+            {
+                return ObjectRotation.Handle.None;
+            }
+            Vec2 c = new(obj.X, obj.Y);
+            double radius = RotationDialRenderer.RadiusPx / View.Zoom;
+            return ObjectRotation.HitTest(
+                c, ObjectRotation.StoredAngle(obj, spec), spec, radius, levelPt,
+                ringTolerance: RotationDialRenderer.RingTolerancePx / View.Zoom,
+                knobTolerance: RotationDialRenderer.KnobTolerancePx / View.Zoom);
+        }
+
+        // Writes the object's new angle from a dial drag: free (whole degrees) unless Alt is held, which
+        // snaps to the spec's step (15°).
+        private static void ApplyRotation(LevelObject obj, RotationSpec spec, Vec2 levelPt, KeyModifiers mods)
+        {
+            bool snap = mods.HasFlag(KeyModifiers.Alt);
+            double angle = ObjectRotation.AngleFromPoint(new Vec2(obj.X, obj.Y), levelPt, spec, snap);
+            obj.SetAttr(spec.AttributeName, ObjectRotation.Format(angle));
+        }
+
+        // Applies the active rail drag to the grab: sliding moves the hook (object x/y) and its offset
+        // together so the rail stays put; resizing an end rewrites moveLength (and moveOffset for the near
+        // end). All constrained by GrabRail so the hook never leaves the rail.
+        private void ApplyRailDrag(LevelObject grab, GrabRail.Geometry g, Vec2 levelPt)
+        {
+            switch (_railDrag)
+            {
+                case GrabRail.Handle.SlideHook:
+                    (double hookAxis, double offset) = GrabRail.SlideHook(g, levelPt);
+                    if (g.Vertical)
+                    {
+                        grab.Y = (int)Math.Round(hookAxis);
+                    }
+                    else
+                    {
+                        grab.X = (int)Math.Round(hookAxis);
+                    }
+                    grab.SetAttr("moveOffset", Whole(offset));
+                    break;
+                case GrabRail.Handle.ResizeEnd:
+                    grab.SetAttr("moveLength", Whole(GrabRail.ResizeEnd(g, levelPt)));
+                    break;
+                case GrabRail.Handle.ResizeStart:
+                    (double offA, double length) = GrabRail.ResizeStart(g, levelPt);
+                    grab.SetAttr("moveOffset", Whole(offA));
+                    grab.SetAttr("moveLength", Whole(length));
+                    break;
+                case GrabRail.Handle.MoveBar:
+                case GrabRail.Handle.None:
+                default:
+                    break;
+            }
+        }
+
+        private static string Whole(double value)
+        {
+            return ((int)Math.Round(value)).ToString(CultureInfo.InvariantCulture);
+        }
+
+        // The cursor for a rail handle: a horizontal rail end/hook reads as a horizontal resize, a vertical
+        // one as a vertical resize (the hook slides along the same axis). The bar keeps the default arrow -
+        // it is still draggable to move the whole grab, but a move cursor over the whole rail is noisy.
+        private Cursor CursorForHandle(GrabRail.Handle handle)
+        {
+            return handle switch
+            {
+                GrabRail.Handle.ResizeStart or GrabRail.Handle.ResizeEnd or GrabRail.Handle.SlideHook =>
+                    SelectedObject is { } s && GrabRail.Vertical(s) ? VResizeCursor : ResizeCursor,
+                GrabRail.Handle.MoveBar => Cursor.Default,
+                GrabRail.Handle.None => Cursor.Default,
+                _ => Cursor.Default,
+            };
+        }
+
+        // Updates the hook hover state, repainting only on a change so the highlight art swaps in/out.
+        private void SetHookHovered(bool hovered)
+        {
+            if (_hookHovered != hovered)
+            {
+                _hookHovered = hovered;
+                InvalidateVisual();
+            }
+        }
+
+        private void SetDialKnobHovered(bool hovered)
+        {
+            if (_dialKnobHovered != hovered)
+            {
+                _dialKnobHovered = hovered;
+                InvalidateVisual();
+            }
+        }
+
+        private static int IndexOf(IReadOnlyList<LevelObject> objects, LevelObject target)
+        {
+            for (int i = 0; i < objects.Count; i++)
+            {
+                if (Equals(objects[i], target))
+                {
+                    return i;
+                }
+            }
+            return -1;
+        }
+
+        private List<LevelBounds> BuildHitBounds(LevelDocument doc)
+        {
+            // Clicking uses the same box that the selection marquee draws (trimmed art + 25%).
+            List<LevelBounds> list = [];
+            foreach (LevelObject o in doc.Objects)
+            {
+                list.Add(Sprites is null
+                    ? new LevelBounds(o.X - 8, o.Y - 8, 16, 16)
+                    : LevelSceneRenderer.SelectionBounds(Sprites, o, ActiveCandySkin, ActiveOmNomSupport, doc.NightLevel));
+            }
+            return list;
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerPressed(PointerPressedEventArgs e)
+        {
+            base.OnPointerPressed(e);
+            LevelDocument? doc = Document;
+            if (doc is null)
+            {
+                return;
+            }
+
+            if (e.GetCurrentPoint(this).Properties.IsMiddleButtonPressed)
+            {
+                _panning = true;
+                _panLast = e.GetPosition(this);
+                e.Pointer.Capture(this);
+                return;
+            }
+            if (!e.GetCurrentPoint(this).Properties.IsLeftButtonPressed)
+            {
+                return;
+            }
+
+            Point p = e.GetPosition(this);
+            Vec2 levelPt = View.ScreenToLevel(new Vec2(p.X, p.Y));
+
+            // Grabbing the auto-catch ring resizes the radius; it takes priority over object hit-testing
+            // (the ring can sit over other objects) but not over a middle-button pan.
+            if (OnRadiusEdge(levelPt))
+            {
+                BeginDocumentEdit?.Invoke();
+                _resizingRadius = true;
+                e.Pointer.Capture(this);
+                return;
+            }
+
+            // Grabbing the selected movable grab's rail: an end cap resizes, the hook slides, the bar moves
+            // the whole grab. Takes priority over hit-testing so the rail wins over anything beneath it.
+            GrabRail.Handle handle = HitRail(levelPt);
+            switch (handle)
+            {
+                case GrabRail.Handle.ResizeStart:
+                case GrabRail.Handle.ResizeEnd:
+                case GrabRail.Handle.SlideHook:
+                    BeginDocumentEdit?.Invoke();
+                    _railDrag = handle;
+                    e.Pointer.Capture(this);
+                    return;
+                case GrabRail.Handle.MoveBar:
+                    BeginDocumentEdit?.Invoke();
+                    _dragOffset = levelPt - new Vec2(SelectedObject!.X, SelectedObject.Y);
+                    _dragging = true;
+                    e.Pointer.Capture(this);
+                    return;
+                case GrabRail.Handle.None:
+                default:
+                    break;
+            }
+
+            // Grabbing the selected object's rotation dial (knob or ring) rotates it; takes priority over
+            // object hit-testing so the dial wins over anything beneath it.
+            if (HitRotationDial(levelPt) != ObjectRotation.Handle.None
+                && SelectedObject is { } rotObj && RotationTable.For(rotObj.Type) is { } rotSpec)
+            {
+                BeginDocumentEdit?.Invoke();
+                _rotating = true;
+                ApplyRotation(rotObj, rotSpec, levelPt, e.KeyModifiers);
+                SelectedObjectMoved?.Invoke();
+                InvalidateVisual();
+                e.Pointer.Capture(this);
+                return;
+            }
+
+            List<LevelBounds> bounds = BuildHitBounds(doc);
+
+            // Double-click toggles the lock. ClickCount keeps climbing (3, 4, ...) while clicking in the
+            // same spot rather than resetting, so every even count counts as a double-click — otherwise a
+            // second double-click at the same position would be missed unless the cursor moved.
+            if (e.ClickCount % 2 == 0)
+            {
+                if (LockedObject is { } current)
+                {
+                    ToggleLock?.Invoke(current);
+                    return;
+                }
+                int dh = HitTester.Topmost(bounds, levelPt, -1);
+                ToggleLock?.Invoke(dh >= 0 ? doc.Objects[dh] : null);
+                return;
+            }
+
+            // While an object is locked, only it is interactive — clicks never fall through to other objects.
+            if (LockedObject is { } locked)
+            {
+                int li = IndexOf(doc.Objects, locked);
+                if (li >= 0 && bounds[li].Contains(levelPt))
+                {
+                    SelectedObject = locked;
+                    BeginDocumentEdit?.Invoke();
+                    _dragOffset = levelPt - new Vec2(locked.X, locked.Y);
+                    _dragging = true;
+                    e.Pointer.Capture(this);
+                }
+                return;
+            }
+
+            int after = _lastHitIndex >= 0 && _lastHitIndex < bounds.Count
+                        && bounds[_lastHitIndex].Contains(levelPt) ? _lastHitIndex : -1;
+            int hit = HitTester.Topmost(bounds, levelPt, after);
+            _lastHitIndex = hit;
+
+            if (hit < 0)
+            {
+                SelectedObject = null;
+                _panning = true;
+                _panLast = p;
+                e.Pointer.Capture(this);
+                return;
+            }
+
+            LevelObject obj = doc.Objects[hit];
+            SelectedObject = obj;
+            BeginDocumentEdit?.Invoke();
+            _dragOffset = levelPt - new Vec2(obj.X, obj.Y);
+            _dragging = true;
+            e.Pointer.Capture(this);
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerMoved(PointerEventArgs e)
+        {
+            base.OnPointerMoved(e);
+            Point p = e.GetPosition(this);
+            Vec2 levelPt = View.ScreenToLevel(new Vec2(p.X, p.Y));
+
+            if (_resizingRadius && SelectedObject is { } g && RadiusRing.Of(g) is { } ring)
+            {
+                double r = GrabRadius.FromDrag(new Vec2(g.X, g.Y), levelPt);
+                g.SetAttr(ring.Attr, ((int)Math.Round(r)).ToString(CultureInfo.InvariantCulture));
+                SelectedObjectMoved?.Invoke();
+                InvalidateVisual();
+                return;
+            }
+
+            if (_railDrag != GrabRail.Handle.None && SelectedObject is { } rg && GrabRail.Of(rg) is { } rail)
+            {
+                ApplyRailDrag(rg, rail, levelPt);
+                SelectedObjectMoved?.Invoke();
+                InvalidateVisual();
+                return;
+            }
+
+            if (_rotating && SelectedObject is { } rotObj && RotationTable.For(rotObj.Type) is { } rotSpec)
+            {
+                ApplyRotation(rotObj, rotSpec, levelPt, e.KeyModifiers);
+                SelectedObjectMoved?.Invoke();
+                InvalidateVisual();
+                return;
+            }
+
+            if (_panning)
+            {
+                ScrollBy(_panLast.X - p.X, _panLast.Y - p.Y);
+                _panLast = p;
+                return;
+            }
+
+            if (!_dragging || SelectedObject is not { } selected)
+            {
+                // Reflect the affordance under the cursor so ring resize / rail edit are discoverable, and
+                // light up the hook when it's hovered.
+                GrabRail.Handle handle = HitRail(levelPt);
+                SetHookHovered(handle == GrabRail.Handle.SlideHook);
+                ObjectRotation.Handle dial = HitRotationDial(levelPt);
+                SetDialKnobHovered(dial == ObjectRotation.Handle.Knob);
+                Cursor = dial != ObjectRotation.Handle.None ? new Cursor(StandardCursorType.Hand)
+                    : OnRadiusEdge(levelPt) ? ResizeCursor : CursorForHandle(handle);
+                return;
+            }
+
+            (int gx, int gy) = Snap(levelPt - _dragOffset);
+            selected.X = gx;
+            selected.Y = gy;
+            SelectedObjectMoved?.Invoke();
+            InvalidateVisual();
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerReleased(PointerReleasedEventArgs e)
+        {
+            base.OnPointerReleased(e);
+            EndPointerGesture();
+            e.Pointer.Capture(null);
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerCaptureLost(PointerCaptureLostEventArgs e)
+        {
+            base.OnPointerCaptureLost(e);
+            EndPointerGesture();
+        }
+
+        private void EndPointerGesture()
+        {
+            // Capture loss (including the release path's own Capture(null)) can fire with nothing in
+            // progress; skip the resets and completion callback unless a gesture is actually active.
+            bool gestureActive = _dragging || _panning || _resizingRadius
+                || _railDrag != GrabRail.Handle.None || _rotating || _hookHovered;
+            if (!gestureActive)
+            {
+                return;
+            }
+
+            bool editedDocument = _dragging || _resizingRadius
+                || _railDrag != GrabRail.Handle.None || _rotating;
+            _dragging = false;
+            _panning = false;
+            _resizingRadius = false;
+            _railDrag = GrabRail.Handle.None;
+            _rotating = false;
+            if (editedDocument)
+            {
+                CompleteDocumentEdit?.Invoke();
+            }
+            // Letting go ends the "grabbed" look; a fresh hover re-lights it if the cursor is on the hook.
+            SetHookHovered(false);
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerExited(PointerEventArgs e)
+        {
+            base.OnPointerExited(e);
+            SetHookHovered(false); // don't leave the hook lit when the cursor leaves the canvas
+            SetDialKnobHovered(false); // nor the rotation knob
+        }
+
+        /// <inheritdoc />
+        protected override void OnPointerWheelChanged(PointerWheelEventArgs e)
+        {
+            base.OnPointerWheelChanged(e);
+            const double wheelPixels = 48;
+            if (e.KeyModifiers.HasFlag(KeyModifiers.Control))
+            {
+                double factor = e.Delta.Y > 0 ? 1.1 : 1 / 1.1;
+                ZoomBy(factor, e.GetPosition(this));
+            }
+            else
+            {
+                double deltaX = e.Delta.X * -wheelPixels;
+                double deltaY = e.Delta.Y * -wheelPixels;
+                if (e.KeyModifiers.HasFlag(KeyModifiers.Shift) && Math.Abs(deltaX) < double.Epsilon)
+                {
+                    deltaX = deltaY;
+                    deltaY = 0;
+                }
+                ScrollBy(deltaX, deltaY);
+            }
+            e.Handled = true;
+        }
+
+        private void Canvas_Pinch(object? sender, PinchEventArgs e)
+        {
+            double factor = ViewNavigation.PinchScaleToZoomFactor(_lastPinchScale, e.Scale);
+            _lastPinchScale = e.Scale;
+            ZoomBy(factor, e.ScaleOrigin);
+            e.Handled = true;
+        }
+
+        private void Canvas_PinchEnded(object? sender, PinchEndedEventArgs e)
+        {
+            _lastPinchScale = 1;
+            e.Handled = true;
+        }
+
+        private void Canvas_TouchPadMagnify(object? sender, PointerDeltaEventArgs e)
+        {
+            double delta = Math.Abs(e.Delta.Y) > double.Epsilon ? e.Delta.Y : e.Delta.X;
+            ZoomBy(ViewNavigation.MagnifyDeltaToZoomFactor(delta), e.GetPosition(this));
+            e.Handled = true;
+        }
+    }
+}
